+++ Changelog.md
@@ -1,3 +1,5 @@
 ## Version 0.1.5
 
 - Change `waitSessionReady` default value to true.
+- Add a waldo-auth binary to setup `~/.waldo/profile.yml` with the user's token.<br>
+  It can be used with `npm exec waldo-auth [TOKEN]`